% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adamreview-package.R
\docType{package}
\name{adamreview-package}
\alias{adamreview}
\alias{adamreview-package}
\title{adamreview: ADaM Dataset Explorer: Review CDISC ADaM Datasets in a Shiny App}
\description{
The ADaM Dataset Explorer is a small golem-based Shiny application for exploring CDISC ADaM clinical trial datasets (data preview, summaries by variable type and interactive charts), comparing two datasets with 'diffdf'. Built as a portfolio project to demonstrate clean, tested Shiny engineering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shreya-sharma20/clinical-data-review}
}

}
\author{
\strong{Maintainer}: Shreya Sharma \email{sshreya319@gmail.com}

Authors:
\itemize{
  \item Shreya Sharma \email{sshreya319@gmail.com}
}

}
\keyword{internal}
