% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the adam-review Shiny application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the case of \code{shinyAppDir}
you can place code in \code{global.R}.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host",
"quiet", "display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment about
the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or \code{"disable"}.
The default, \code{NULL}, respects any previous call to
\code{\link[shiny:enableBookmarking]{shiny::enableBookmarking()}}, and otherwise defaults to \code{"disable"}.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the request.}

\item{...}{Arguments to pass to \code{golem::get_golem_options()}. See \code{?golem::get_golem_options}
for more details.}
}
\description{
Run the adam-review Shiny application
}
